🍿  Move web folder
