+++ src/export/index.ts
@@ -4,7 +4,6 @@ export * from './tex';
 export * from './pdf';
 export * from './notebook';
 export * from './jupyter-book';
-export * as web from './web';
 export * from './model';
 export * from './utils';
 export * from './multiple';
