+++ src/index.ts
@@ -1,3 +1,4 @@
 export * from './session';
 export * from './models';
 export * from './export';
+export * as web from './web';
