+++ src/web/transforms/footnotes.ts
